+++ lib/Models/registerCatalogMembers.ts
@@ -1,5 +1,7 @@
+import ArcGisCatalogGroup from "./ArcGisCatalogGroup";
 import ArcGisFeatureServerCatalogGroup from "./ArcGisFeatureServerCatalogGroup";
 import ArcGisFeatureServerCatalogItem from "./ArcGisFeatureServerCatalogItem";
+import ArcGisMapServerCatalogGroup from "./ArcGisMapServerCatalogGroup";
 import ArcGisMapServerCatalogItem from "./ArcGisMapServerCatalogItem";
 import BingMapsCatalogItem from "./BingMapsCatalogItem";
 import CartoMapCatalogItem from "./CartoMapCatalogItem";
@@ -39,9 +41,14 @@ export default function registerCatalogMembers() {
   CatalogMemberFactory.register(GeoJsonCatalogItem.type, GeoJsonCatalogItem);
   CatalogMemberFactory.register(CsvCatalogItem.type, CsvCatalogItem);
   CatalogMemberFactory.register(CzmlCatalogItem.type, CzmlCatalogItem);
+  CatalogMemberFactory.register(ArcGisCatalogGroup.type, ArcGisCatalogGroup);
   CatalogMemberFactory.register(
     ArcGisMapServerCatalogItem.type,
     ArcGisMapServerCatalogItem
+  );
+  CatalogMemberFactory.register(
+    ArcGisMapServerCatalogGroup.type,
+    ArcGisMapServerCatalogGroup
   );
   CatalogMemberFactory.register(
     ArcGisFeatureServerCatalogItem.type,
@@ -133,6 +140,11 @@ export default function registerCatalogMembers() {
     true
   );
   createCatalogItemFromUrl.register(
+    matchesUrl(/\/arcgis\/rest\/.*\/MapServer(\/.*)?$/i),
+    ArcGisMapServerCatalogGroup.type,
+    true
+  );
+  createCatalogItemFromUrl.register(
     matchesUrl(/\/arcgis\/rest\/.*\/FeatureServer\/\d+\b/i),
     ArcGisFeatureServerCatalogItem.type,
     true
@@ -148,8 +160,18 @@ export default function registerCatalogMembers() {
     true
   );
   createCatalogItemFromUrl.register(
+    matchesUrl(/\/arcgis\/rest\//i),
+    ArcGisCatalogGroup.type,
+    true
+  );
+  createCatalogItemFromUrl.register(
     matchesUrl(/\/rest\/.*\/MapServer\/\d+\b/i),
     ArcGisMapServerCatalogItem.type,
+    true
+  );
+  createCatalogItemFromUrl.register(
+    matchesUrl(/\/rest\/.*\/MapServer(\/.*)?$/i),
+    ArcGisMapServerCatalogGroup.type,
     true
   );
   createCatalogItemFromUrl.register(
@@ -181,9 +203,15 @@ export default function registerCatalogMembers() {
   );
   createCatalogItemFromUrl.register(
     s => true,
+    ArcGisMapServerCatalogGroup.type,
+    true
+  );
+  createCatalogItemFromUrl.register(
+    s => true,
     ArcGisFeatureServerCatalogItem.type,
     true
   );
+  createCatalogItemFromUrl.register(s => true, ArcGisCatalogGroup.type, true);
   createCatalogItemFromUrl.register(
     s => true,
     ArcGisFeatureServerCatalogGroup.type,